(maint) Update gems as far as possible
This commit updates all of our gem dependencies to latest versions
whenever possible, and as new as possible to avoid dependency resolution
conflicts besides that. This list plays nicely together.

It also removes a few puppet-lint plugins that have since been included
in puppet-lint itself. This update required a couple of test changes:

* When the `legacy_facts` plugin was incorporated into core puppet-lint,
they changed the message it emits when a legacy fact is detected to
include the name of the fact. This commit updates the test to match the
new message.
* There was also a new check added that ensures there is only a single
space before an arrow `=>` when there is only a single parameter in a
resource. We had a resource with a second param commented out, where we
never adjusted the spacing for the one that was left. This now failed
the linter, so this commit updates the spacing to make it happy, since
that was in the `goodsyntax` test.